AK+Everywhere: Change int to size_t in JsonObject and JsonArray

+++ Userland/Libraries/LibKeyboard/CharacterMapFile.cpp
@@ -74,7 +74,7 @@ Vector<u32> CharacterMapFile::read_map(const JsonObject& json, const String& nam
     buffer.resize(CHAR_MAP_SIZE);
 
     auto map_arr = json.get(name).as_array();
-    for (int i = 0; i < map_arr.size(); i++) {
+    for (size_t i = 0; i < map_arr.size(); i++) {
         auto key_value = map_arr.at(i).as_string();
         if (key_value.length() == 0) {
             buffer[i] = 0;
